moved UserMessage to single file
